backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
    "asyncpg>=0.29.0",
    "uvicorn[standard]>=0.32.0",
    "loguru>=0.7.3",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]
